+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - 新增对 `.drawio.svg`, `.drawio.html`, `.drawio.xml` 三种可编辑格式的导出支持 | Added export support for `.drawio.svg`, `.drawio.html`, `.drawio.xml` in three editable formats.
 - 优化 URL 参数解析方案 | Optimized URL parameter resolution schemes.
+- 将文件扩展名 `.drawio.svg`, `.drawio.html`, `.drawio.xml` 更改为 `.svg`, `.html`, `.xml` | Change the `.drawio.svg`, `.drawio.html`, `.drawio.xml` filename extensions to `.svg`, `.html`, `.xml`.
 
 ## v0.1.2/2022-03-02